pyproject.toml
[project]
name = "open-llms-play-pokemon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">= 3.10, < 3.12"
dependencies = [
    "dotenv>=0.9.9",
    "openai>=1.82.0",
    "pillow>=11.2.1",
    "pyboy>=2.6.0",
    "dspy-ai>=2.6.27",
]

[dependency-groups]
dev = [
    "modal>=1.0.2",
 "pytest>=8.4.0",
 "ruff>=0.8.0",
 "pyright>=1.1.400",
 "mlflow>=2.22.1",
]


[tool.ruff]
target-version = "py310"
line-length = 88
extend-exclude = ["build", "game"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by formatter
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
include = ["open-llms-play-pokemon", "server"]
exclude = ["build", "game", ".venv"]
pythonVersion = "3.10"
typeCheckingMode = "standard"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = true
reportUnusedClass = true
reportUnusedFunction = true
reportUnusedVariable = true
